Minor compatibility changes #21

+++ include/nil/blueprint/transpiler/templates/permutation_argument.hpp
@@ -56,7 +56,7 @@ library modular_permutation_argument_$TEST_NAME${
         uint256 h = 1;
         uint256 g = 1;
         
-        for(uint8 i = 0; i < permutation_size;){
+        for(uint256 i = 0; i < permutation_size;){
             uint256 tmp = addmod(
                 gamma, 
                 basic_marshalling.get_uint256_be(
